Added first enemy, HUD
Added the first enemy, a slugger. added an enemy package, added enemy
class, moved some code to MapObject to reduce redundancy, added the HUD
so player can see health and fireball level. Made Fireballs reverse if
fired to the left.
TODO: fix the fireball animation glitch (not showing properly) fix enemy
disappearing when player leaves area and returns, tweak the camera
system.

+++ Andreas game/src/com/charleswomble/TileMap/TileMap.java
@@ -125,14 +125,14 @@ public class TileMap
 		return tileSize;
 	}
 	
-	public int getx()
+	public double getx()
 	{
-		return (int)x;
+		return x;
 	}
 	
-	public int gety()
+	public double gety()
 	{
-		return (int)y;
+		return y;
 	}
 	
 	public int getWidth()
